move logic to support class
